Update text field style
