+++ forui/lib/src/widgets/text_field/text_field.dart
@@ -9,6 +9,7 @@ import 'package:meta/meta.dart';
 
 import 'package:forui/forui.dart';
 
+part 'text_field_state_style.dart';
 part 'text_field_style.dart';
 part 'text_form_field.dart';
